move alert store to state-folder
